pyproject.toml
[project]
name = "learn-llm-rag"
version = "0.1.0"
description = "Learn to implement LLM and RAG application"
readme = "README.md"
authors = [
    { name = "Rajeev Karamcheti", email = "rkaramc@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langchain-ollama>=0.2.1",
    "python-dotenv>=1.0.1",
]

[project.scripts]
rag = "learn_llm_rag:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "pip>=24.3.1",
]
